requirements.txt
asgiref==3.2.3
astroid==2.4.2
bcrypt==3.2.0
boto3==1.16.51
botocore==1.19.51
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
cryptography==3.3.1
dateutils==0.6.12
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.7
django-cache-headers==0.4
django-cors-headers==3.2.1
django-environ==0.4.5
django-filter==2.2.0
django-jsonfield==1.3.1
django-ses==1.0.3
djangorestframework==3.11.0
djangorestframework-datatables==0.5.1
environ==1.0
et-xmlfile==1.0.1
future==0.18.2
gunicorn==20.0.4
idna==2.8
image==1.5.33
isort==5.7.0
jdcal==1.4.1
jmespath==0.10.0
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
# mysqlclient @ file:///F:/mysqlclient-1.4.6-cp39-cp39-win_amd64.whl
odfpy==1.4.0
openpyxl==3.0.2
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0
paypalrestsdk==1.13.1
Pillow==8.1.0
pycparser==2.19
pylint==2.6.0
pyOpenSSL==20.0.1
python-dateutil==2.6.0
pytz==2019.3
PyYAML==5.2
requests==2.22.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.0
tablib==0.14.0
toml==0.10.2
urllib3==1.25.8
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
